requirements.txt
pymongo
yfinance
langchain
langchain-openai
python-dotenv
